Move from hardcoded list to frontmatter
Example use from the Buildroot doc:
```md
---
title: The Buildroot Toolchain
permalink: build/buildroot.html
summary: Information and usage of the Buildroot tool, used to compile dahliaOS
---
## Buildroot
```
`title` is used for the breadcrumb.
`permalink` is auto generated when needed.
`summary` is the description, and I'm not sure where it's used.

Signed-off-by: Blake <bleonard252@rocketmail.com>